\name{GIMLET-package}
\alias{GIMLET-package}
\alias{GIMLET}
\docType{package}
\title{
\packageTitle{GIMLET}
}
\description{
\packageDescription{GIMLET}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GIMLET}
\packageIndices{GIMLET}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{GIMLET}

Maintainer: \packageMaintainer{GIMLET}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
