% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimlet.R
\name{gimlet}
\alias{gimlet}
\title{Genome-wide identification of modulators with local energy test statistics}
\usage{
gimlet(X, Y, Z, r = 1, delta = 0.3, perm = 1000, type = 1)
}
\arguments{
\item{X}{data matrix of X}

\item{Y}{data matrix of Y}

\item{Z}{data matrix of Z}

\item{r}{number of samples in each of two groups with high and low values of Z}

\item{delta}{tuning parameter controlling the smoothness of the local estimator}

\item{perm}{number of permutations for calculating p-values}
}
\value{
a list consisting of
\item{pi}{the approximation of pi}
\item{n}{the number of trials}
\item{hits}{the number of hits}
\item{modulator}{Selected value of Z}
\item{ldcor}{Local estimator of distance correlation between X and Y given selected values of Z}
\item{global.score}{Test statistic evaluating whether the correlation difference between X and Y depends on Z}
\item{each.p.value}{p-value in statistical test under the null hypothesis H_0: the local estimator between X and Y given Z is equal to 0}
\item{global.p.value}{p-value in statistical test under the null hypothesis H_0: the local estimator betwee X and Y does not depend on Z}
\item{null.global.score}{Test statistics under the null hypothesis}
\item{null.each.ldcor}{Local estimators under the null hypothesis}
\item{corrected.each.p.value}{Corrected p-value in statistical test under the null hypothesis H_0: the local estimator between X and Y given Z is equal to 0 with the distribution tail approximation}
\item{corrected.global.p.value}{Corrected p-value in statistical test under the null hypothesis H_0: the local estimator betwee X and Y does not depend on Z  with the distribution tail approximation}
}
\description{
Genome-wide identification of modulators with local energy test statistics
}
\details{
This function calculates the local estimator of distance correlation between random vectors X and Y of arbitorary dimension given a specific value of another random vector Z and tests the statistical siginificance of the correlation differences between two different groups of Z.
}
\examples{
data <- sim(500); result <- gimlet(data$x,data$y,data$z,1,0.3,1000)
}
\author{
Teppei Shimamura & Yusuke Matsui
}
\references{
Teppei Shimamura, Yusuke Matsui, Taisuke Kajino, Satoshi Ito, Takashi Takahashi and Satoru Miyano. Genome-wide identification of biological modulators with local energy statistics. submitted.
}

